agents/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# OpenAI SDK (upgraded for Responses API support with tools)
openai==2.6.1

# Utilities
python-dotenv==1.0.0

# Image processing
Pillow==10.1.0

# HTML parsing
beautifulsoup4==4.12.2

# Corpus management
PyYAML==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Browser automation for iframe validation
playwright==1.40.0

# Development
black==23.11.0
ruff==0.1.6

